+++ tests/render_order.test.ts
@@ -1,12 +1,14 @@
 import { assertStrictEquals as equals } from "../deps/assert.ts";
-import { getSite } from "./utils.ts";
+import { build, getSite } from "./utils.ts";
 
 Deno.test("render order property", async () => {
   const site = getSite({
     src: "render_order",
   });
 
-  await site.build();
+  site.addEventListener("afterRender", () => false);
+
+  await build(site);
 
   const pages = site.pages;